Move test util modules to src/utils/test
