+++ frontend/src/utils/test/mergeResolvers.js
